… src/performance/cdn/cdn-optimizer.js
… src/performance/database/query-optimizer.js
… src/performance/performance-suite.js
… src/performance/virtualization/virtual-scroller.js
… src/performance/wasm/power-calculations.js
… src/performance/workers/analytics-worker.js
… src/performance/workers/worker-manager.js
… tests/unit/performance-suite.test.js